+++ Calculator/src/main/java/ru/nsu/sartakov/operations/AddOperator.java
@@ -1,29 +1,30 @@
-package Operations;
+package ru.nsu.sartakov.operations;
 
 import java.util.EmptyStackException;
 import java.util.Stack;
 
-public class SubOperator implements Operation{
+public class AddOperator implements Operation{
     /**
+     *
      * @param stack - the input data presented as stack
-     * @return subtraction of the last two numbers
+     * @return the sum of the last two elements in stack
      */
     @Override
     public double calculate(Stack<Double> stack) {
-        if (stack.size() > 1) {
+        if (stack.size() > 1 ) {
             double a = stack.pop();
             double b = stack.pop();
-            return a - b;
+            return a + b;
         } else {
             throw new EmptyStackException();
         }
     }
 
     /**
-     * @return the representation of an operation
+     * @return the operation char
      */
     @Override
     public String getRepresentation() {
-        return "-";
+        return "+";
     }
 }
