+++ Calculator/src/main/java/ru/nsu/sartakov/operations/DivOperator.java
@@ -1,30 +1,31 @@
-package Operations;
+package ru.nsu.sartakov.operations;
 
 import java.util.EmptyStackException;
 import java.util.Stack;
 
-public class AddOperator implements Operation{
+public class DivOperator implements Operation{
     /**
      *
      * @param stack - the input data presented as stack
-     * @return the sum of the last two elements in stack
+     * @return the result of the last two numbers division
      */
     @Override
     public double calculate(Stack<Double> stack) {
-        if (stack.size() > 1 ) {
+        if (stack.size() > 1) {
             double a = stack.pop();
             double b = stack.pop();
-            return a + b;
+            return a / b;
         } else {
             throw new EmptyStackException();
         }
     }
 
     /**
-     * @return the operation char
+     *
+     * @return the representation of an operation
      */
     @Override
     public String getRepresentation() {
-        return "+";
+        return "/";
     }
 }
