+++ Calculator/src/main/java/ru/nsu/sartakov/operations/PowOperator.java
@@ -1,23 +1,24 @@
-package Operations;
+package ru.nsu.sartakov.operations;
 
 import java.util.EmptyStackException;
 import java.util.Stack;
 
-public class DivOperator implements Operation{
+public class PowOperator implements Operation {
     /**
      *
      * @param stack - the input data presented as stack
-     * @return the result of the last two numbers division
+     * @return the result of powering the last number for the previous number
      */
     @Override
     public double calculate(Stack<Double> stack) {
         if (stack.size() > 1) {
             double a = stack.pop();
             double b = stack.pop();
-            return a / b;
+            return Math.pow(a, b);
         } else {
             throw new EmptyStackException();
         }
+
     }
 
     /**
@@ -26,6 +27,6 @@ public class DivOperator implements Operation{
      */
     @Override
     public String getRepresentation() {
-        return "/";
+        return "pow";
     }
 }
